+++ src/selection_sort.cpp
@@ -7,9 +7,9 @@ namespace algorithm
     {
         void selectionSort(std::vector<int>& v)
         {
-            for (int i = 0; i < v.size(); ++i)
+            for (size_t i = 0; i < v.size(); ++i)
             {
-                for (int j = i + 1; j < v.size(); ++j)
+                for (size_t j = i + 1; j < v.size(); ++j)
                 {
                     if (v[i] > v[j])
                     {
